// apps/notification-ws/src/types.ts
export type SpaceEvent =
  | {
    type: "JOIN_REQUEST";
    payload: {
      adminUserId: string;
      requester: {
        userId: string;
        userName: string;
        email: string;
        inviteId: string;
      };
    };
  }
  | {
    type: "JOIN_APPROVED";
    payload: {
      userId: string;
      spaceSlug: string;
      spaceName: string;
    };
  }
  | {
    type: "USER_BLOCKED";
    payload: {
      userId: string;
      spaceName: string;
    };
  };